Move implementation into package
