+++ hypothesis-python/src/hypothesis/extra/lark.py
@@ -42,7 +42,7 @@ import attr
 import lark
 from lark.grammar import NonTerminal, Terminal
 
-import hypothesis._strategies as st
+import hypothesis.strategies.internal.core as st
 from hypothesis.errors import InvalidArgument
 from hypothesis.internal.compat import getfullargspec, string_types
 from hypothesis.internal.conjecture.utils import calc_label_from_name
