+++ hypothesis-python/tests/cover/test_type_lookup.py
@@ -23,7 +23,6 @@ import pytest
 
 import hypothesis.strategies as st
 from hypothesis import given, infer
-from hypothesis._strategies import _strategies
 from hypothesis.errors import (
     HypothesisDeprecationWarning,
     InvalidArgument,
@@ -32,6 +31,7 @@ from hypothesis.errors import (
 from hypothesis.internal.compat import PY2, integer_types
 from hypothesis.searchstrategy import types
 from hypothesis.searchstrategy.types import _global_type_lookup
+from hypothesis.strategies.internal.core import _strategies
 from tests.common.utils import checks_deprecated_behaviour
 
 # Build a set of all types output by core strategies
